fixed single line html problem
regex has a problem recognizing multiple results when they are on the
same line. By forcing a newline, the regex pattern recognizer will grab
all the results. This also let me find that the [To Parent Directory]
was recognized as a match. I've hard coded a removal of that result.
hope it's not too big of a problem later on.