Bugfix: Crash while saving when hosts was disabled and enabled together
When the HostsFile was disabled (hosts.disabled file exists and no hosts file), saving would lead the program to create a new hosts file. If the user wanted to disable the HostsFile the again, the program would crash.

Now before saving the program checks if there exists a hosts.disabled file and if yes, tells the user that he/she needs to enable the HostsFile before editing entries.